reorg validator directory
