+++ adit/core/validators.py
@@ -68,11 +68,3 @@ def validate_series_number(value: str) -> None:
             raise ValueError()
     except ValueError:
         raise ValidationError(f"Invalid series number: {value}")
-
-
-def validate_pseudonym(pseudonym: str) -> None:
-    no_backslash_char_validator(pseudonym)
-    no_control_chars_validator(pseudonym)
-    no_wildcard_chars_validator(pseudonym)
-    if len(pseudonym) > 64:
-        raise ValidationError("Pseudonym string too long (max 64 characters).")
